final_requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.1
djangorestframework==3.14.0
et-xmlfile==1.1.0
idna==3.4
mysql-connector-python==8.0.33
mysqlclient==2.2.0
openpyxl==3.1.2
paywix==2.0.2
protobuf==3.20.3
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.6.2
tzdata==2023.3
urllib3==1.26.6
whitenoise==6.4.0
